requirements.txt
h5py
numpy
scipy
pytorch
transformers
perceiver-io